backend/app/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
spotipy==2.24.0
python-dotenv==1.0.1
gunicorn==23.0.0
aiofiles==24.0.0